addressed the comments
