+++ client/errors.go
@@ -1,0 +1,5 @@
+package client
+
+import "fmt"
+
+var ErrMarketNotProvided = fmt.Errorf("provide a market")
