--- client/private_endpoints.go
@@ -1,3 +1,0 @@
-package client
-
-// TODO
