+++ go.mod
@@ -4,11 +4,20 @@ go 1.21.4
 
 require github.com/stretchr/testify v1.8.4
 
-require golang.org/x/net v0.17.0 // indirect
+require (
+	github.com/btcsuite/btcd/btcec/v2 v2.2.0 // indirect
+	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.0.1 // indirect
+	github.com/holiman/uint256 v1.2.4 // indirect
+	golang.org/x/crypto v0.17.0 // indirect
+	golang.org/x/net v0.18.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/ethereum/go-ethereum v1.13.12
 	github.com/gorilla/websocket v1.5.1
+	github.com/huandu/xstrings v1.4.0
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
